requirements.txt
BeautifulSoup==3.2.1
Django==1.7.1
SQLAlchemy==0.9.8
argparse==1.2.1
boto==2.34.0
csvkit==0.9.0
dbf==0.94.003
jdcal==1.0
latimes-calculate==0.1.8
numpy==1.9.0
openpyxl==2.1.2
python-dateutil==2.2
python-elections==1.1.0
six==1.8.0
wsgiref==0.1.2
xlrd==0.9.3
